ElectionPilot/requirements.txt

bcrypt>=5.0.0
flask>=3.1.2
flask-limiter>=3.13
flask-mail>=0.10.0
flask-migrate>=4.1.0
flask-sqlalchemy>=3.1.1
flask-wtf>=1.2.2
numpy>=2.3.3
opencv-python>=4.11.0.86
pandas>=2.3.2
pillow>=11.3.0
psycopg2-binary>=2.9.10
pytest>=8.4.2
python-dotenv>=1.1.1
reportlab>=4.4.4
openpyxl>=3.1.5
sqlalchemy>=2.0.0
werkzeug>=3.0.0
requests>=2.32.0
bcrypt
flask
flask-limiter
flask-mail
flask-migrate
flask-sqlalchemy
flask-wtf
numpy
opencv-python
openpyxl
pandas
pillow
psycopg2-binary
pytest
python-dotenv
reportlab
requests
sqlalchemy
werkzeug
